Move token-related functions to their own module
bacon: test
Resolves: OKTA-89850
