+++ lib/token.js
@@ -1,0 +1,489 @@
+/* globals FRAME_ID */
+/*jshint maxcomplexity:15 */
+
+var http          = require('./http');
+var util          = require('./util');
+var Q             = require('q');
+var sdkCrypto     = require('./crypto');
+var AuthSdkError  = require('./errors/AuthSdkError');
+var OAuthError    = require('./errors/OAuthError');
+
+function getWellKnown(sdk) {
+  return http.get(sdk, sdk.options.url + '/.well-known/openid-configuration');
+}
+
+function validateClaims(claims, iss, aud) {
+  if (!claims || !iss || !aud) {
+    throw new AuthSdkError('The jwt, iss, and aud arguments are all required');
+  }
+
+  var now = Math.floor(new Date().getTime()/1000);
+
+  if (claims.iss !== iss) {
+    throw new AuthSdkError('The issuer [' + claims.iss + '] ' +
+      'does not match [' + iss + ']');
+  }
+
+  if (claims.aud !== aud) {
+    throw new AuthSdkError('The audience [' + claims.aud + '] ' +
+      'does not match [' + aud + ']');
+  }
+
+  if (claims.iat > claims.exp) {
+    throw new AuthSdkError('The JWT expired before it was issued');
+  }
+
+  if (now > claims.exp) {
+    throw new AuthSdkError('The JWT expired and is no longer valid');
+  }
+
+  if (claims.iat > now) {
+    throw new AuthSdkError('The JWT was issued in the future');
+  }
+}
+
+function decodeIdToken(idToken) { /* jshint ignore:line */
+  var jwt = idToken.split('.');
+  var decodedToken;
+
+  try {
+    decodedToken = {
+      header: JSON.parse(util.base64UrlToString(jwt[0])),
+      payload: JSON.parse(util.base64UrlToString(jwt[1])),
+      signature: jwt[2]
+    };
+  } catch(e) {
+    throw new AuthSdkError('Malformed idToken');
+  }
+
+  return decodedToken;
+}
+
+function verifyIdToken(sdk, idToken, options) { /* jshint ignore:line */
+  options = options || {};
+
+  if (!sdk.features.isTokenVerifySupported()) {
+    return Q.reject(new AuthSdkError('This browser doesn\'t support crypto.subtle'));
+  }
+
+  function isExpired(jwtExp) {
+    var expirationTime;
+    if (options.expirationTime || options.expirationTime === 0) {
+      expirationTime = options.expirationTime;
+    } else {
+      expirationTime = Math.floor(Date.now()/1000);
+    }
+    if (jwtExp &&
+        jwtExp > expirationTime) {
+      return true;
+    }
+  }
+
+  function hasAudience(jwtAudience) {
+    if (!options.audience) {
+      return true;
+    }
+    var audiences = Array.isArray(options.audience) ? options.audience : [options.audience];
+    var jwtAudiences = Array.isArray(jwtAudience) ? jwtAudience : [jwtAudience];
+    var ai = audiences.length;
+    while (ai--) {
+      var aud = audiences[ai];
+      if (jwtAudiences.indexOf(aud) !== -1) {
+        return true;
+      }
+    }
+  }
+
+  return getWellKnown(sdk)
+    .then(function(res) {
+      return http.get(sdk, res['jwks_uri']);
+    })
+    .then(function(res) {
+      var key = res.keys[0];
+      return sdkCrypto.verifyToken(idToken, key);
+    })
+    .then(function(res) {
+      if (!res) {
+        return false;
+      }
+      var jwt = sdk.idToken.decode(idToken);
+
+      if (isExpired(jwt.payload.exp)) {
+        return false;
+      }
+
+      if (!hasAudience(jwt.payload.aud)) {
+        return false;
+      }
+
+      if (options.issuer &&
+          options.issuer !== jwt.payload.iss) {
+        return false;
+      }
+
+      return true;
+    });
+}
+
+function refreshIdToken(sdk, opts) { /* jshint ignore:line */
+  opts = opts || {};
+  opts.display = null;
+  opts.prompt = 'none';
+  return getIdToken(sdk, opts);
+}
+
+function loadFrame(src, iframeId) {
+  if (typeof iframeId === 'undefined') {
+    return;
+  }
+
+  var iframe = document.getElementById(iframeId);
+  if (iframe) {
+    return iframe;
+  }
+  
+  iframe = document.createElement('iframe');
+  iframe.setAttribute('id', iframeId);
+  iframe.style.display = 'none';
+  iframe.src = src;
+
+  return document.body.appendChild(iframe);
+}
+
+function loadPopup(src, options) {
+  var title = options.popupTitle || 'External Identity Provider User Authentication';
+  var appearance = 'toolbar=no, scrollbars=yes, resizable=yes, ' +
+    'top=100, left=500, width=600, height=600';
+  return window.open(src, title, appearance);
+}
+
+function addListener(eventTarget, name, fn) {
+    if (eventTarget.addEventListener) {
+      eventTarget.addEventListener(name, fn);
+    } else {
+      eventTarget.attachEvent('on' + name, fn);
+    }
+  }
+
+function removeListener(eventTarget, name, fn) {
+  if (eventTarget.removeEventListener) {
+    eventTarget.removeEventListener(name, fn);
+  } else {
+    eventTarget.detachEvent('on' + name, fn);
+  }
+}
+
+function addPostMessageListener(sdk, timeout) {
+  var deferred = Q.defer();
+
+  function responseHandler(e) {
+    if (!e.data || e.origin !== sdk.options.url) {
+      return;
+    }
+    deferred.resolve(e.data);
+  }
+
+  addListener(window, 'message', responseHandler);
+
+  return deferred.promise.timeout(timeout || 120000, new AuthSdkError('OAuth flow timed out'))
+    .fin(function() {
+      removeListener(window, 'message', responseHandler);
+    });
+}
+
+function addFragmentListener(sdk, windowEl, timeout) {
+  var deferred = Q.defer();
+
+  // Predefine regexs for parsing hash
+  var plus2space = /\+/g;
+  var paramSplit = /([^&=]+)=?([^&]*)/g;
+
+  function hashToObject(hash) {
+    // Remove the leading hash
+    var fragment = hash.substring(1);
+
+    var obj = {};
+
+    // Loop until we have no more params
+    var param;
+    while (true) {
+      param = paramSplit.exec(fragment);
+      if (!param) { break; }
+
+      var key = param[1];
+      var value = param[2];
+
+      // id_token should remain base64url encoded
+      if (key === 'id_token') {
+        obj[key] = value;
+      } else {
+        obj[key] = decodeURIComponent(value.replace(plus2space, ' '));
+      }
+    }
+    return obj;
+  }
+
+  function hashChangeHandler() {
+    /*
+      We are only able to access window.location.hash on a window
+      that has the same domain. A try/catch is necessary because
+      there's no other way to determine that the popup is in
+      another domain. When we try to access a window on another 
+      domain, an error is thrown.
+    */
+    try {
+      if (windowEl &&
+          windowEl.location &&
+          windowEl.location.hash) {
+        deferred.resolve(hashToObject(windowEl.location.hash));
+      } else if (windowEl && !windowEl.closed) {
+        setTimeout(hashChangeHandler, 500);
+      }
+    } catch (err) {
+      setTimeout(hashChangeHandler, 500);
+    }
+  }
+
+  hashChangeHandler();
+
+  return deferred.promise.timeout(timeout || 120000, new AuthSdkError('OAuth flow timed out'));
+}
+
+/*
+ * Retrieve an idToken from an Okta or a third party idp
+ * 
+ * Two main flows:
+ *
+ *  1) Exchange a sessionToken for an idToken
+ * 
+ *    Required:
+ *      clientId: passed via the OktaAuth constructor or into getIdToken
+ *      sessionToken: 'yourtoken'
+ *
+ *    Optional:
+ *      redirectUri: defaults to window.location.href
+ *      scopes: defaults to ['openid', 'email']
+ *
+ *    Forced:
+ *      prompt: 'none'
+ *      responseMode: 'okta_post_message'
+ *      display: undefined
+ *
+ *  2) Get an idToken from an idp
+ *
+ *    Required:
+ *      clientId: passed via the OktaAuth constructor or into getIdToken
+ *
+ *    Optional:
+ *      redirectUri: defaults to window.location.href
+ *      scopes: defaults to ['openid', 'email']
+ *      idp: defaults to Okta as an idp
+ *      prompt: no default. Pass 'none' to throw an error if user is not signed in
+ *
+ *    Forced:
+ *      display: 'popup'
+ *
+ *  Only common optional params shown. Any OAuth parameters not explicitly forced are available to override
+ *
+ * @param {Object} oauthOptions
+ * @param {String} [oauthOptions.clientId] ID of this client
+ * @param {String} [oauthOptions.redirectUri] URI that the iframe or popup will go to once authenticated
+ * @param {String[]} [oauthOptions.scopes] OAuth 2.0 scopes to request (openid must be specified)
+ * @param {String} [oauthOptions.idp] ID of an external IdP to use for user authentication
+ * @param {String} [oauthOptions.sessionToken] Bootstrap Session Token returned by the Okta Authentication API
+ * @param {String} [oauthOptions.prompt] Determines whether the Okta login will be displayed on failure.
+ *                                       Use 'none' to prevent this behavior
+ *
+ * @param {Object} options
+ * @param {Integer} [options.timeout] Time in ms before the flow is automatically terminated. Defaults to 120000
+ * @param {String} [options.popupTitle] Title dispayed in the popup.
+ *                                      Defaults to 'External Identity Provider User Authentication'
+ */
+function getIdToken(sdk, oauthOptions, options) { /* jshint ignore:line */
+  if (!oauthOptions) {
+    oauthOptions = {};
+  }
+
+  if (!options) {
+    options = {};
+  }
+
+  // Default OAuth query params
+  var oauthParams = {
+    clientId: sdk.options.clientId,
+    redirectUri: sdk.options.redirectUri || window.location.href,
+    responseType: 'id_token',
+    responseMode: 'okta_post_message',
+    state: util.genRandomString(64),
+    nonce: util.genRandomString(64),
+    scope: ['openid', 'email']
+  };
+
+  // Add user-provided options
+  util.extend(oauthParams, oauthOptions);
+
+  // Start overriding any options that don't make sense
+  var sessionTokenOverrides = {
+    prompt: 'none',
+    responseMode: 'okta_post_message',
+    display: null
+  };
+
+  var idpOverrides = {
+    display: 'popup'
+  };
+
+  if (oauthOptions.sessionToken) {
+    util.extend(oauthParams, sessionTokenOverrides);
+  } else if (oauthOptions.idp) {
+    util.extend(oauthParams, idpOverrides);
+  }
+
+  // Quick validation
+  if (!oauthParams.clientId) {
+    throw new AuthSdkError('A clientId must be specified in the OktaAuth constructor to get an idToken');
+  }
+
+  // Convert our params to their actual OAuth equivalents
+  var oauthQueryHash = util.removeNils({
+    'client_id': oauthParams.clientId,
+    'redirect_uri': oauthParams.redirectUri,
+    'response_type': oauthParams.responseType,
+    'response_mode': oauthParams.responseMode,
+    'state': oauthParams.state,
+    'nonce': oauthParams.nonce,
+    'prompt': oauthParams.prompt,
+    'display': oauthParams.display,
+    'sessionToken': oauthParams.sessionToken,
+    'idp': oauthParams.idp
+  });
+
+  if (oauthParams.scope.indexOf('openid') !== -1) {
+    oauthQueryHash.scope = oauthParams.scope.join(' ');
+  } else {
+    throw new AuthSdkError('openid scope must be specified in the scope argument');
+  }
+
+  // Use the query params to build the authorize url
+  var requestUrl = sdk.options.url + '/oauth2/v1/authorize' + util.toQueryParams(oauthQueryHash);
+
+  // Determine the flow type
+  var flowType;
+  if (oauthParams.sessionToken || oauthParams.display === null) {
+    flowType = 'IFRAME';
+  } else if (oauthParams.display === 'popup') {
+    flowType = 'POPUP';
+  } else {
+    flowType = 'IMPLICIT';
+  }
+
+  function handleOAuthResponse(res) {
+    if (res['error'] || res['error_description']) {
+      throw new OAuthError(res['error'], res['error_description']);
+
+    } else if (res['id_token']) {
+      if (res.state !== oauthParams.state) {
+        throw new AuthSdkError('OAuth flow response state doesn\'t match request state');
+      }
+      if (res.nonce !== oauthParams.nonce) {
+        throw new AuthSdkError('OAuth flow response nonce doesn\'t match request nonce');
+      }
+      var jwt = sdk.idToken.decode(res['id_token']);
+      validateClaims(jwt.payload, sdk.options.url, oauthParams.clientId);
+      return {
+        idToken: res['id_token'],
+        claims: jwt.payload
+      };
+
+    } else {
+      throw new AuthSdkError('Unable to parse OAuth flow response');
+    }
+  }
+
+  function getOrigin(url) {
+    var originRegex = /^(https?\:\/\/)?([^:\/?#]*(?:\:[0-9]+)?)/;
+    return originRegex.exec(url)[0];
+  }
+
+  // Execute the flow type
+  switch (flowType) {
+    case 'IFRAME':
+      var iframePromise = addPostMessageListener(sdk, options.timeout);
+      var iframeEl = loadFrame(requestUrl, FRAME_ID);
+      return iframePromise
+        .then(handleOAuthResponse)
+        .fin(function() {
+          if (document.body.contains(iframeEl)) {
+            iframeEl.parentElement.removeChild(iframeEl);
+          }
+        });
+
+    case 'POPUP':
+      var popupPromise;
+
+      // Add listener on postMessage before window creation, so
+      // postMessage isn't triggered before we're listening
+      if (oauthParams.responseMode === 'okta_post_message') {
+        if (!sdk.features.isPopupPostMessageSupported()) {
+          return Q.reject(new AuthSdkError('This browser doesn\'t have full postMessage support'));
+        }
+        popupPromise = addPostMessageListener(sdk, options.timeout);
+      }
+
+      // Create the window
+      var windowOptions = {
+        popupTitle: options.popupTitle
+      };
+      var windowEl = loadPopup(requestUrl, windowOptions);
+
+      // Poll until we get a valid hash fragment
+      if (oauthParams.responseMode === 'fragment') {
+        var windowOrigin = getOrigin(sdk.idToken.authorize._getLocationHref());
+        var redirectUriOrigin = getOrigin(oauthParams.redirectUri);
+        if (windowOrigin !== redirectUriOrigin) {
+          return Q.reject(new AuthSdkError('Using fragment, the redirectUri origin (' + redirectUriOrigin +
+            ') must match the origin of this page (' + windowOrigin + ')'));
+        }
+        popupPromise = addFragmentListener(sdk, windowEl, options.timeout);
+      }
+
+      // Both postMessage and fragment require a poll to see if the popup closed
+      var popupDeferred = Q.defer();
+      function hasClosed(win) {
+        if (win.closed) {
+          popupDeferred.reject(new AuthSdkError('Unable to parse OAuth flow response'));
+        }
+      }
+      var closePoller = setInterval(function() {
+        hasClosed(windowEl);
+      }, 500);
+
+      // Proxy the promise results into the deferred
+      popupPromise
+      .then(function(res) {
+        popupDeferred.resolve(res);
+      })
+      .fail(function(err) {
+        popupDeferred.reject(err);
+      });
+
+      return popupDeferred.promise
+        .then(handleOAuthResponse)
+        .fin(function() {
+          if (!windowEl.closed) {
+            clearInterval(closePoller);
+            windowEl.close();
+          }
+        });
+
+    default:
+      return Q.reject(new AuthSdkError('The full page redirect flow is not supported'));
+  }
+}
+
+module.exports = {
+	getIdToken: getIdToken,
+	refreshIdToken: refreshIdToken,
+	decodeIdToken: decodeIdToken,
+	verifyIdToken: verifyIdToken
+};
