+++ package.json
@@ -54,6 +54,7 @@
   },
   "okta-auth-js": {
     "STATE_TOKEN_COOKIE_NAME": "oktaStateToken",
-    "DEFAULT_POLLING_DELAY": 500
+    "DEFAULT_POLLING_DELAY": 500,
+    "FRAME_ID": "okta-oauth-helper-frame"
   }
 }
